requirements.txt
numpy
pandas
wheel
setuptools
hydra-core 
google-cloud-compute
